datatype change

+++ Sadie.Database/Models/Players/PlayerRelationship.cs
@@ -5,9 +5,9 @@ namespace Sadie.Database.Models.Players;
 public class PlayerRelationship
 {
     public int Id { get; init; }
-    public int OriginPlayerId { get; init; }
+    public long OriginPlayerId { get; init; }
     public Player? OriginPlayer { get; init; }
-    public int TargetPlayerId { get; init; }
+    public long TargetPlayerId { get; init; }
     public Player? TargetPlayer { get; init; }
     public PlayerRelationshipType TypeId { get; set; }
 }
